Move sortByFields to parent scan builder
